+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement `Action` for `&Action` and `&mut Action`.
 
+### Changed
+
+- `Fold` requires 2 generic params: `State` and `Heap`.
+- `Fold::fold` and inline fold will provide `input`.
+
 ## [0.1.0] - 2024-12-29
 
 ### Added